use extra argument on scheme_make_fd_output_port
This is necessary on the computer I'm using right now (x86), but it was
not necessary on my normal computer (x86_64).  They are both up-to-date
Arch Linux installations, so they should have the same version of
Racket.  I originally tried using the extra argument and it failed on my
normal computer... but maybe I'm just crazy.  It works fine on my wife's
x86_64 Ubuntu install.  I guess if there are problems with it later I'll
fix it then...