add temporary backspace hack
To fit in with the long history of the backspace key sending the delete
character instead of the backspace character, supposedly started by
Linus Torvalds.  Such confusion.  This will go away when I add some sort
of configuration for what gets sent for what keys.  Don't let the st
terminal people see that I added this!